requirements.txt
trakt.py>=2.12.0, < 4.5.0

click>=6.0, < 8.2.0
pytz>=2016.6
six>=1.10.0
rich>=12.0
